+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/core/fees/TransactionPriceCalculator.java
@@ -37,10 +37,10 @@ public interface TransactionPriceCalculator {
       if (transaction.getType().equals(TransactionType.FRONTIER)) {
         return transaction.getGasPrice();
       }
-      final Wei gasPremium =
-          Wei.of((BigInteger) transaction.getGasPremium().orElseThrow().getValue());
+      final Wei minerFee =
+          Wei.of((BigInteger) transaction.getMinerFee().orElseThrow().getValue());
       final Wei feeCap = Wei.of((BigInteger) transaction.getFeeCap().orElseThrow().getValue());
-      Wei price = gasPremium.add(baseFee);
+      Wei price = minerFee.add(baseFee);
       if (price.compareTo(feeCap) > 0) {
         price = feeCap;
       }