+++ ethereum/core/src/test/java/org/hyperledger/besu/ethereum/mainnet/MainnetTransactionValidatorTest.java
@@ -287,7 +287,7 @@ public class MainnetTransactionValidatorTest {
     final Transaction transaction =
         new TransactionTestFixture()
             .type(TransactionType.EIP1559)
-            .gasPremium(Optional.of(Wei.of(3)))
+            .minerFee(Optional.of(Wei.of(3)))
             .feeCap(Optional.of(Wei.of(6)))
             .gasLimit(21000)
             .chainId(Optional.empty())
@@ -321,7 +321,7 @@ public class MainnetTransactionValidatorTest {
     final Transaction transaction =
         new TransactionTestFixture()
             .type(TransactionType.EIP1559)
-            .gasPremium(Optional.of(Wei.of(1)))
+            .minerFee(Optional.of(Wei.of(1)))
             .feeCap(Optional.of(Wei.of(1)))
             .chainId(Optional.empty())
             .createTransaction(senderKeys);
@@ -345,7 +345,7 @@ public class MainnetTransactionValidatorTest {
             defaultGoQuorumCompatibilityMode);
     final Transaction transaction =
         new TransactionTestFixture()
-            .gasPremium(Optional.of(Wei.of(1)))
+            .minerFee(Optional.of(Wei.of(1)))
             .feeCap(Optional.of(Wei.of(1)))
             .chainId(Optional.empty())
             .createTransaction(senderKeys);